Add class support.
New ecmascript classes are supported now.
This is the simplest case.
Maybe some edge cases won't work yet.

'use strict' is required for node 4 to be able to handle classes.
As we're only using classes in the tests, we will not throw an linting
error for 'use strict'.
The same is true for the spread operator. This is only supported in
node 4 if the `--harmony` flag is enabled.